requirements.txt
bcrypt==4.0.1
bson==0.5.10
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.3
zipp==3.13.0
